requirements.txt
streamlit==1.44.1
yfinance==0.2.54
pandas==2.2.3
numpy==2.2.4
plotly==6.0.1
python-dotenv==1.0.1
google-generativeai==0.8.4
scipy==1.15.2